+++ packages/core/src/rules/index.ts
@@ -9,7 +9,11 @@ import {
   contractMissingRule,
   contractOutOfScopeRule,
 } from "./contractRules.js";
-import { commitTrailerForbiddenRule, commitTrailerMissingRule } from "./commitTrailers.js";
+import {
+  commitAiDisclosedRule,
+  commitTrailerForbiddenRule,
+  commitTrailerMissingRule,
+} from "./commitTrailers.js";
 import { contentUnavailableRule } from "./contentUnavailable.js";
 import { highRiskPathRule } from "./highRiskPath.js";
 import { packageScriptDriftRule } from "./packageScripts.js";
@@ -32,6 +36,7 @@ export const builtInRules: Rule[] = [
   packageScriptDriftRule,
   commitTrailerMissingRule,
   commitTrailerForbiddenRule,
+  commitAiDisclosedRule,
   workflowPermissionEscalationRule,
   workflowTriggerRemovedRule,
   workflowDangerousPatternRule,